refactor: use createClassNames from utils
