+++ packages/react-instantsearch-dom/src/components/ClearRefinements.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import classNames from 'classnames';
-import createClassNames from '../core/createClassNames';
+import { createClassNames } from '../core/utils';
 import translatable from '../core/translatable';
 
 const cx = createClassNames('ClearRefinements');
