+++ packages/react-instantsearch-dom/src/components/CurrentRefinements.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import createClassNames from '../core/createClassNames';
+import { createClassNames } from '../core/utils';
 import translatable from '../core/translatable';
 
 const cx = createClassNames('CurrentRefinements');
