+++ packages/react-instantsearch-dom/src/components/HierarchicalMenu.js
@@ -1,7 +1,7 @@
 import { pick } from 'lodash';
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import createClassNames from '../core/createClassNames';
+import { createClassNames } from '../core/utils';
 import translatable from '../core/translatable';
 import List from './List';
 import Link from './Link';
