+++ packages/react-instantsearch-dom/src/components/Highlight.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import createClassNames from '../core/createClassNames';
+import { createClassNames } from '../core/utils';
 import Highlighter from './Highlighter';
 
 const cx = createClassNames('Highlight');
