+++ packages/react-instantsearch-dom/src/components/NumericMenu.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import translatable from '../core/translatable';
-import createClassNames from '../core/createClassNames';
+import { createClassNames } from '../core/utils';
 import List from './List';
 
 const cx = createClassNames('NumericMenu');
