+++ packages/react-instantsearch-dom/src/components/Panel.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import createClassNames from '../core/createClassNames';
+import { createClassNames } from '../core/utils';
 
 const cx = createClassNames('Panel');
